feat(header): convert to typescript
